+++ controller/controlplane/controller_reconciler_utils.go
@@ -2,8 +2,10 @@ package controlplane
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
+	"github.com/samber/lo"
 	certificatesv1 "k8s.io/api/certificates/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -11,6 +13,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kcfgcontrolplane "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/controlplane"
+	operatorv1alpha1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v1alpha1"
 	operatorv1beta1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v1beta1"
 	operatorv2alpha1 "github.com/kong/kubernetes-configuration/v2/api/gateway-operator/v2alpha1"
 
@@ -44,7 +47,7 @@ func (r *Reconciler) ensureIsMarkedScheduled(
 }
 
 // ensureDataPlaneStatus ensures that the dataplane is in the correct state
-// to carry on with the controlplane deployments reconciliation.
+// to carry on with the controlplane reconciliation.
 // Information about the missing dataplane is stored in the controlplane status.
 func (r *Reconciler) ensureDataPlaneStatus(
 	cp *ControlPlane,
@@ -119,4 +122,127 @@ func (r *Reconciler) ensureAdminMTLSCertificateSecret(
 		r.Client,
 		matchingLabels,
 	)
+}
+
+func (r *Reconciler) validateWatchNamespaceGrants(
+	ctx context.Context,
+	cp *ControlPlane,
+) ([]string, error) {
+	if cp.Spec.WatchNamespaces == nil {
+		return nil, errors.New("spec.watchNamespaces cannot be empty")
+	}
+
+	switch cp.Spec.WatchNamespaces.Type {
+	// NOTE: We currentlty do not require any ReferenceGrants or other permission
+	// granting resources for the "All" case.
+	case operatorv2alpha1.WatchNamespacesTypeAll:
+		return nil, nil
+	// No special permissions are required to watch the controlplane's own namespace.
+	case operatorv2alpha1.WatchNamespacesTypeOwn:
+		return []string{cp.Namespace}, nil
+	case operatorv2alpha1.WatchNamespacesTypeList:
+		var nsList []string
+		for _, ns := range cp.Spec.WatchNamespaces.List {
+			if err := ensureWatchNamespaceGrantsForNamespace(ctx, r.Client, cp, ns); err != nil {
+				return nsList, err
+			}
+			nsList = append(nsList, ns)
+		}
+		// Add ControlPlane's own namespace as it will add it anyway because
+		// that's where the default "publish service" exists.
+		// We add it here as we do not require a ReferenceGrant for own namespace
+		// so there's no validation whether a grant exists.
+		nsList = append(nsList, cp.Namespace)
+
+		return nsList, nil
+	default:
+		return nil, fmt.Errorf("unexpected watchNamespaces.type: %q", cp.Spec.WatchNamespaces.Type)
+	}
+}
+
+// +kubebuilder:rbac:groups=gateway-operator.konghq.com,resources=watchnamespacegrants,verbs=list
+
+// ensureWatchNamespaceGrantsForNamespace ensures that a WatchNamespaceGrant exists for the
+// given namespace and ControlPlane.
+// It returns an error if a WatchNamespaceGrant is missing.
+func ensureWatchNamespaceGrantsForNamespace(
+	ctx context.Context,
+	cl client.Client,
+	cp *ControlPlane,
+	ns string,
+) error {
+	var grants operatorv1alpha1.WatchNamespaceGrantList
+	if err := cl.List(ctx, &grants, client.InNamespace(ns)); err != nil {
+		return fmt.Errorf("failed listing WatchNamespaceGrants in namespace %s: %w", ns, err)
+	}
+	for _, grant := range grants.Items {
+		if watchNamespaceGrantContainsControlPlaneFrom(grant, cp) {
+			// return nil if there is one grant allows the CP to watch the namespace.
+			return nil
+		}
+	}
+	return fmt.Errorf("WatchNamespaceGrant in Namespace %s to ControlPlane in Namespace %s not found", ns, cp.Namespace)
+}
+
+func watchNamespaceGrantContainsControlPlaneFrom(
+	grant operatorv1alpha1.WatchNamespaceGrant,
+	cp *ControlPlane,
+) bool {
+	for _, from := range grant.Spec.From {
+		if from.Group == operatorv2alpha1.SchemeGroupVersion.Group &&
+			from.Kind == "ControlPlane" &&
+			from.Namespace == cp.Namespace {
+			return true
+		}
+	}
+	return false
+}
+
+// validateWatchNamespaces validates that the operator's watch namespaces are compatible
+// with the ControlPlane's watch namespaces configuration.
+func validateWatchNamespaces(
+	cp *ControlPlane,
+	watchNamespaces []string,
+) error {
+	if cp.Spec.WatchNamespaces == nil {
+		return nil
+	}
+
+	// If ControlPlane is configured to watch all namespaces, operator should
+	// not be configured to watch any specific namespaces: it should have the watchNamespaces
+	// list empty, indicating that it will watch all namespaces.
+	switch cp.Spec.WatchNamespaces.Type {
+	case operatorv2alpha1.WatchNamespacesTypeAll:
+		if len(watchNamespaces) > 0 {
+			return fmt.Errorf(
+				"ControlPlane's watchNamespaces is set to 'All', but operator is only allowed on: %v",
+				watchNamespaces,
+			)
+		}
+
+	case operatorv2alpha1.WatchNamespacesTypeOwn:
+		// NOTE: In case the operator Pod does not have watch namespaces flag/environment variable set
+		// to include the ControlPlane's namespace, it will not be able to watch
+		// the ControlPlane's own namespace, which is required for the operator to function properly.
+		if len(watchNamespaces) > 0 && !lo.Contains(watchNamespaces, cp.Namespace) {
+			return fmt.Errorf(
+				"ControlPlane's watchNamespaces is set to 'Own' (current ControlPlane namespace: %v), but operator is only allowed on: %v",
+				cp.Namespace, watchNamespaces,
+			)
+		}
+
+	case operatorv2alpha1.WatchNamespacesTypeList:
+		if len(watchNamespaces) == 0 {
+			return nil
+		}
+
+		if !lo.Every(watchNamespaces, cp.Spec.WatchNamespaces.List) {
+			return fmt.Errorf(
+				"ControlPlane's watchNamespaces requests %v, but operator is only allowed on: %v",
+				cp.Spec.WatchNamespaces.List, watchNamespaces,
+			)
+		}
+	}
+
+	return nil
 }